+++ src/http/query.rs
@@ -3,9 +3,9 @@ use std::fmt;
 
 use std::collections::HashSet;
 
-#[deriving(Decodable)]
+#[derive(RustcDecodable)]
 pub struct Request {
-    pub scalars: Vec<uint>,
+    pub scalars: Vec<usize>,
 }
 
 impl clone::Clone for Request {
@@ -24,13 +24,13 @@ impl fmt::Show for Request {
     }
 }
 
-#[deriving(Encodable)]
+#[derive(RustcEncodable)]
 pub struct Response {
     pub scalars: Vec<ScalarResult>,
 }
 
-#[deriving(Encodable)]
+#[derive(RustcEncodable)]
 pub struct ScalarResult {
-    pub scalar: uint,
-    pub found: HashSet<uint>,
+    pub scalar: usize,
+    pub found: HashSet<usize>,
 }
